backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
supabase==2.9.0
httpx==0.27.0
python-dotenv==1.0.0
